pyproject.toml
[project]
name = "provenance_track"
version = "1.2.8"
dependencies = []
requires-python= ">= 3.8"
readme = "README.md"
authors = [
    {name = "Gerard"},
    {email = "gweatherby@uchc.edu"}
]
dynamic = ["license"]

[project.optional-dependencies]
mock = ['keyrings.efile',
    'postgresql-access',
    'pyyaml']

#[project.scripts]
#provenance_track = "provenance_track.main:main"
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

